refactor(store): organize store module into repository/subscription/snapshot subdirectories
Pure module reorganization — no public API changes. Files moved:
- builder, event_store, zero_copy_event_store, repository, replay -> repository/
- subscription, checkpoint -> subscription/
- snapshotting -> snapshot/

Internal `use` paths updated from `super::` to `crate::store::` where needed.
All re-exports preserved in store/mod.rs.
