+++ crates/nexus-store/src/store/mod.rs
@@ -1,11 +1,7 @@
-mod builder;
-mod checkpoint;
-mod event_store;
 mod raw;
-pub(crate) mod replay;
 mod repository;
 #[cfg(feature = "snapshot")]
-mod snapshotting;
+mod snapshot;
 #[allow(
     clippy::module_inception,
     reason = "Store<S> is the primary public type of the store module"
@@ -13,18 +9,15 @@ mod snapshotting;
 mod store;
 mod stream;
 mod subscription;
-mod zero_copy_event_store;
 
-#[cfg(feature = "snapshot")]
-pub use builder::WithSnapshot;
-pub use builder::{NeedsCodec, NoSnapshot, RepositoryBuilder};
-pub use checkpoint::CheckpointStore;
-pub use event_store::EventStore;
 pub use raw::RawEventStore;
-pub use repository::Repository;
 #[cfg(feature = "snapshot")]
-pub use snapshotting::Snapshotting;
+pub use repository::WithSnapshot;
+pub use repository::{
+    EventStore, NeedsCodec, NoSnapshot, Repository, RepositoryBuilder, ZeroCopyEventStore,
+};
+#[cfg(feature = "snapshot")]
+pub use snapshot::Snapshotting;
 pub use store::Store;
 pub use stream::{EventStream, EventStreamExt};
-pub use subscription::Subscription;
-pub use zero_copy_event_store::ZeroCopyEventStore;
+pub use subscription::{CheckpointStore, Subscription};
