+++ crates/nexus-store/src/store/snapshot/mod.rs
@@ -1,0 +1,3 @@
+mod snapshotting;
+
+pub use snapshotting::Snapshotting;
