+++ crates/nexus-store/src/store/subscription/mod.rs
@@ -1,0 +1,9 @@
+mod checkpoint;
+#[allow(
+    clippy::module_inception,
+    reason = "Subscription trait is the primary public type of the subscription module"
+)]
+mod subscription;
+
+pub use checkpoint::CheckpointStore;
+pub use subscription::Subscription;
